refactor: 라우트 전용 폴더 네이밍 규칙 정리(#278)
- 라우트 바로 아래 유틸 폴더를 `_utils`로 정리
- feature 내부 `_utils`를 `utils`로 변경
- applications, dashboard, landing 관련 import 경로 갱신
- README 프로젝트 구조와 네이밍 규칙 문서 업데이트
