+++ app/(protected)/applications/_utils/preview.ts
