+++ app/_components/landing/utils/content.ts
